Fix some errors and bugs, repair logging

+++ src/includes/sme/stage/FileUtils.hxx
@@ -9,7 +9,7 @@
 
 namespace SME::Util
 {
-    s8 getCharacterID(JUTGamePad *controller = nullptr);
+    s32 getCharacterID(JUTGamePad *controller = nullptr);
     const char *getStageName(TApplication *gpApplication);
     void *loadArchive(char *path, JKRHeap *heap);
 }
